Added path.create path error handling Now path.create:
n. tokenizes paths
n. dynamically checks each token joined to the previous, the current working path, to see if the path exists
n. if not it creates the current working path
n. returns all valid paths (which is all of them by this point)

All paths in lib and tests are OS independent